Refactor create page, extrude functions into useCreateProjectProcess
